reorganize dirs a bit
