+++ src/backend/backend.ts
@@ -2,6 +2,7 @@
 // =======
 //
 // - This is the electron main thread.
+// - Entry point is defined in _package.json_
 // - Initial config parsing is here for convenience.
 // - Window and tray handling **must be** here though
 // - You access backend functionality through electron ipc sync messages
@@ -9,7 +10,7 @@
 // - Some messages handled by the backend: `getConfig`, `resizeWindow`, `centerWindow` etc.
 //
 import { app } from 'electron';
-import { log } from '../interim/log';
+import { log } from '../shared/log';
 import { initBackend } from './initBackend';
 
 app
