+++ src/backend/initGlobalShortcuts.ts
@@ -1,7 +1,7 @@
 import { globalShortcut } from 'electron';
-import { log } from '../interim/log';
-import { getConfig } from '../interim/getConfig';
-import { sendMessageAtBackend } from './modules/messages/sendMessageAtBackend';
+import { log } from '../shared/log';
+import { getConfig } from '../shared/getConfig';
+import { sendMessageAtBackend } from './messaging/sendMessageAtBackend';
 
 const initialized = false;
