+++ src/renderer/runtime/parsing/getSearchableText.ts
