+++ src/renderer/runtime/parsing/postParsingHooks/processXdgDesktopFile.ts
