+++ src/renderer/runtime/sendMessage.ts
@@ -1,5 +1,5 @@
 import { ipcRenderer } from 'electron';
-import { messages } from '../backend/messages';
+import { messages } from '../../backend/messages';
 
 export function sendMessage(messageId: keyof typeof messages, payload?) {
   if (!messages[messageId]) throw new Error('Unknown message id!');
