+++ src/renderer/runtime/setWindowSize.ts
@@ -1,8 +1,8 @@
-import { sendMessage } from '../../interim/sendMessage';
 import { constants } from '../constants';
 import { sharedConfig } from '../shared/sharedConfig';
 import { sharedStore } from '../shared/sharedStore';
 import { $ } from '../utils/dom';
+import { sendMessage } from './sendMessage';
 
 /**
  * Sends a `MSG_RESIZE_WINDOW` message to the backend
