+++ src/shared/utils/readJsonFile.ts
@@ -1,13 +1,14 @@
 import { access, readFile } from 'fs/promises';
 import { log } from '../log';
 
-export async function readJsonFile(fileName) {
+export async function readJsonFile(fileName): Promise<string | null> {
   try {
     await access(fileName);
   } catch {
+    log.error(`No access to file "${fileName}"`);
     return null;
   }
-  let contents = null;
+  let contents: string;
   try {
     contents = await readFile(fileName, 'utf8');
   } catch {
